Add initial ultrawide/surround support
This makes sure the HUD does not expand to a box wider than 16:9
to preserve readability, especially in fast-paced games such as
first person shooters. (Human peripheral vision only goes so far.)

The trick is to set the HUD's root Control node to anchor to
VCenter Wide. This will work as long as the project resolution uses
a 16:9 aspect ratio.